refactoring package structure
