+++ MobileTracker/app/src/main/java/me/myds/g2u/mobiletracker/data/TransactionResult.java
@@ -1,4 +1,4 @@
-package me.myds.g2u.mobiletracker.IconRPC;
+package me.myds.g2u.mobiletracker.data;
 
 import androidx.annotation.NonNull;
